Add files via upload
The pdf file contains a report on the findings through the logistic regression model in R.
The R file contains all the code for the logistic regression.

Conclusion of the analysis:
We can use this model by observing closely the prior education of prisoners and the weeks of experience the prisoners accumulate before deciding on whether the prisoners should be sent back to prison or not. All prisoners should be given financial aid since it lowers their chances of being re-arrested by 40% based on the coefficient. Additionally, using the findings of other variables having an impact on arrests made when individually observed, we should only consider prisoners who have at least completed their high school education, have prior work experience, have been married and have no prior convictions to be enrolled in a program where they are eligible to receive financial aid. After receiving the financial aid for a few weeks, prisoners should be observed carefully to determine whether they can find employment or not. If they cannot find employment after the financial aid has been given to them for a period up to 6 months, they should be re-arrested and sent to prison as they are likely to commit a crime based on this predictive model. If employment is found by the convicts within half a year then according to this model, those convicts have a high chance of remaining clean and not being arrested again. This step-by-step program will ensure the successful transition of former convicts into civilian life.